+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Changelog
   * new: Backend.needs_config
   * new: Track.ids and gpxdo fix --id-from-backend
   * new: gpxdo cp
+  * new: Track.first_different_point()
 
 1.3.2 release 2018-08-19
 ------------------------